requirements.txt
numpy
pandas
opencv-python
tensorflow==2.3.1
keras==2.4.3
scikit-learn
pillow
streamlit